requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.80.0
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.4
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.19.1
typing_extensions==4.4.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
google-cloud-bigquery
google-api-python-client
black